Change homing durability values to float

+++ src/core/DevilDaggersInfo.Core.Wiki/Structs/HomingDamage.cs
@@ -1,3 +1,3 @@
 namespace DevilDaggersInfo.Core.Wiki.Structs;
 
-public readonly record struct HomingDamage(int FromLevel3, int FromLevel4, int FromLevel4Splash, int ToLevel3, int ToLevel4);
+public readonly record struct HomingDamage(int FromLevel3, int FromLevel4, int FromLevel4Splash, float ToLevel3, float ToLevel4);
